WhatsChanged added extension methods
